ROGO-2767 need to delete paper settings when deleting a paper

+++ classes/paperutils.class.php
@@ -785,8 +785,17 @@ class PaperUtils
      * @param mysqli $db
      * @return bool true on success, false otherwise
      */
-    static function complete_delete_paper($id, $db)
+    public static function complete_delete_paper($id, $db)
     {
+        // Delete from paper settings first.
+        $result = $db->prepare('DELETE FROM paper_settings WHERE paperid = ?');
+        $result->bind_param('i', $id);
+        $result->execute();
+        if ($db->errno != 0) {
+            return false;
+        }
+
+        $result->close();
         $result = $db->prepare('DELETE FROM properties WHERE property_id = ?');
         $result->bind_param('i', $id);
         $result->execute();
